requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.135.2
fastapi-openai-compat==1.1.0
filelock==3.25.2
filetype==1.2.0
fsspec==2026.2.0
h11==0.16.0
haystack-ai==2.26.1
haystack-experimental==0.19.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpath-ng==1.8.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_imports==1.2.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==2.29.0
packaging==26.0
posthog==7.9.12
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-ulid==3.1.0
PyYAML==6.0.3
redis==7.3.0
redisvl==0.16.0
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
